requirements.txt
fastapi
uvicorn
cryptography
psycopg2-binary
redis